pyproject.toml
[tool.poetry]
name = "humioapi"
version = "0.8.0"
description = "An unofficial Python library wrapping the official humiolib to provide extra helpers and utility functions"
authors = ["Jostein Haukeli"]
repository = "https://github.com/gwtwod/humioapi"
readme = "README.md"

packages = [
    { include = "humioapi" }
]

[tool.poetry.dependencies]
python = "^3.6.1"
snaptime = "^0.2.4"
structlog = "^20.1.0"
tzlocal = "^2.1"
chardet = "^4.0.0"
python-dotenv = "^0.15.0"
tqdm = "^4.54.1"
pendulum = "^2.1.2"
colorama = "^0.4.4"
pandas = {version = "^1.1.5", optional = true}

[tool.poetry.dev-dependencies]
black = {version = "*", allow-prereleases = true}
pylint = "*"
pre-commit = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-bandit = "*"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
(/(
    \.eggs
  | \.git
  | \.venv
  | _?build
  | dist
)/
  | setup\.py
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
